Package Creation for SentinelOne
