+++ Solutions/SentinelOne/Analytic Rules/SentinelOneAdminLoginNewIP.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Admin login from new location
 description: |
   'Detects admin user login from new location (IP address).'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
