+++ Solutions/SentinelOne/Analytic Rules/SentinelOneAgentUninstalled.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Agent uninstalled from multiple hosts
 description: |
   'Detects when agent was uninstalled from multiple hosts.'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
