+++ Solutions/SentinelOne/Analytic Rules/SentinelOneAlertFromCustomRule.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Alert from custom rule
 description: |
   'Detects when alert from custom rule received.'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
