+++ Solutions/SentinelOne/Analytic Rules/SentinelOneBlacklistHashDeleted.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Blacklist hash deleted
 description: |
   'Detects when blacklist hash was deleted.'
 severity: Medium
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
