+++ Solutions/SentinelOne/Analytic Rules/SentinelOneExclusionAdded.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Exclusion added
 description: |
   'Detects when new exclusion added.'
 severity: Medium
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
