+++ Solutions/SentinelOne/Analytic Rules/SentinelOneMultipleAlertsOnHost.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Multipl alerts on host
 description: |
   'Detects when multiple alerts received from same host.'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
