+++ Solutions/SentinelOne/Analytic Rules/SentinelOneNewAdmin.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - New admin created
 description: |
   'Detects when new admin user is created.'
 severity: Medium
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
