+++ Solutions/SentinelOne/Analytic Rules/SentinelOneRuleDeleted.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Rule deleted
 description: |
   'Detects when a rule was deleted.'
 severity: Medium
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
