+++ Solutions/SentinelOne/Analytic Rules/SentinelOneRuleDisabled.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Rule disabled
 description: |
   'Detects when a rule was disabled.'
 severity: Medium
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
