+++ Solutions/SentinelOne/Analytic Rules/SentinelOneSameCustomRuleHitOnDiffHosts.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - Same custom rule triggered on different hosts
 description: |
   'Detects when same custom rule was triggered on different hosts.'
 severity: High
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
