+++ Solutions/SentinelOne/Analytic Rules/SentinelOneViewAgentPassphrase.yaml
@@ -3,6 +3,7 @@ name: Sentinel One - User viewed agent's passphrase
 description: |
   'Detects when a user viewed agent's passphrase.'
 severity: Medium
+status: Available
 requiredDataConnectors:
   - connectorId: SentinelOne
     dataTypes:
