+++ Solutions/SentinelOne/Data Connectors/SentinelOneSentinelConnector/function.json
