+++ Solutions/SentinelOne/Data Connectors/SentinelOneSentinelConnector/state_manager.py
