+++ Solutions/SentinelOne/Data Connectors/azuredeploy_Connector_SentinelOneAPI_AzureFunction.json
