+++ Solutions/SentinelOne/Data Connectors/host.json
