+++ Solutions/SentinelOne/SolutionMetadata.json
@@ -1,0 +1,15 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-sentinelone",
+	"firstPublishDate": "2022-04-01",
+	"providers": [ "Microsoft " ],
+	"categories": {
+		"domains": [ "Security - Threat Protection" ]
+	},
+	"support": {
+		"name": "Microsoft Corporation",
+		"email": "support@microsoft.com",
+		"tier": "Microsoft",
+		"link": "https://support.microsoft.com"
+	}
+}
